Fix flakey retry tests
